requirements.txt
falcon==1.1.0
gevent==1.2.1
pymongo==3.4.0
transitions==0.2.8
PyYAML==3.11
python-dateutil==2.6.0
requests==2.13.0
